feat: Clean Architectureに沿うようにロジックを分割した
