+++ backend/app/domain/todo/todo.go
@@ -1,0 +1,7 @@
+package todo
+
+type Todo struct {
+	Id        int
+	Title     string
+	Completed bool
+}
